requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-dev.in -o requirements-dev.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.11
    # via
    #   aiohttp-jinja2
    #   textual-dev
    #   textual-serve
aiohttp-jinja2==1.6
    # via textual-serve
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   asyncer
    #   httpx
    #   watchfiles
asgi-lifespan==2.1.0
    # via -r requirements-dev.in
asttokens==3.0.0
    # via stack-data
asyncer==0.0.8
    # via -r requirements-dev.in
attrs==25.3.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
black==25.1.0
    # via -r requirements-dev.in
blinker==1.9.0
    # via flask
brotli==1.1.0
    # via geventhttpclient
bunnet==1.3.0
    # via -r requirements-dev.in
certifi==2025.4.26
    # via
    #   geventhttpclient
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   bunnet
    #   flask
    #   hatch
    #   rich-toolkit
    #   textual-dev
    #   typer
    #   userpath
    #   uvicorn
configargparse==1.7.1
    # via
    #   locust
    #   locust-cloud
coverage==7.8.2
    # via -r requirements-dev.in
decorator==5.2.1
    # via ipython
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via pymongo
executing==2.2.0
    # via stack-data
fastapi-cli==0.0.7
    # via -r requirements-dev.in
filelock==3.18.0
    # via virtualenv
flask==3.1.1
    # via
    #   flask-cors
    #   flask-login
    #   locust
flask-cors==6.0.0
    # via locust
flask-login==0.6.3
    # via locust
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
gevent==25.5.1
    # via
    #   geventhttpclient
    #   locust
    #   locust-cloud
geventhttpclient==2.3.3
    # via locust
greenlet==3.2.3
    # via gevent
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
hatch==1.14.1
    # via -r requirements-dev.in
hatch-requirements-txt==0.4.1
    # via -r requirements-dev.in
hatch-vcs==0.5.0
    # via -r requirements-dev.in
hatchling==1.27.0
    # via
    #   -r requirements-dev.in
    #   hatch
    #   hatch-requirements-txt
    #   hatch-vcs
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via hatch
hyperlink==21.0.0
    # via hatch
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
ipython==9.3.0
    # via -r requirements-dev.in
ipython-pygments-lexers==1.1.1
    # via ipython
itsdangerous==2.2.0
    # via flask
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   aiohttp-jinja2
    #   flask
    #   textual-serve
keyring==25.6.0
    # via hatch
lazy-model==0.2.0
    # via bunnet
linkify-it-py==2.0.3
    # via markdown-it-py
locust==2.37.10
    # via -r requirements-dev.in
locust-cloud==1.23.1
    # via locust
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.7
    # via ipython
mdit-py-plugins==0.4.2
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via
    #   jaraco-classes
    #   jaraco-functools
msgpack==1.1.0
    # via
    #   locust
    #   textual-dev
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via
    #   pre-commit
    #   pyright
packaging==25.0
    # via
    #   black
    #   hatch
    #   hatch-requirements-txt
    #   hatchling
    #   pytest
    #   setuptools-scm
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   hatchling
pexpect==4.9.0
    # via
    #   hatch
    #   ipython
platformdirs==4.3.8
    # via
    #   black
    #   hatch
    #   locust-cloud
    #   textual
    #   virtualenv
pluggy==1.6.0
    # via
    #   hatchling
    #   pytest
pre-commit==4.2.0
    # via pre-commit-uv
pre-commit-uv==4.1.4
    # via -r requirements-dev.in
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via locust
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.11.5
    # via
    #   bunnet
    #   lazy-model
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
    #   rich
pymongo==4.13.0
    # via bunnet
pyright==1.1.401
    # via -r requirements-dev.in
pytest==8.4.0
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
pytest-asyncio==1.0.0
    # via -r requirements-dev.in
python-dotenv==1.1.0
    # via uvicorn
python-engineio==4.12.2
    # via
    #   locust-cloud
    #   python-socketio
python-socketio==5.13.0
    # via locust-cloud
pyyaml==6.0.2
    # via
    #   pre-commit
    #   uvicorn
pyzmq==26.4.0
    # via locust
requests==2.32.3
    # via
    #   locust
    #   python-socketio
rich==14.0.0
    # via
    #   hatch
    #   rich-toolkit
    #   textual
    #   textual-serve
    #   typer
rich-toolkit==0.14.7
    # via fastapi-cli
ruff==0.11.13
    # via -r requirements-dev.in
setuptools==80.9.0
    # via
    #   locust
    #   setuptools-scm
    #   zope-event
    #   zope-interface
setuptools-scm==8.3.1
    # via hatch-vcs
shellingham==1.5.4
    # via
    #   hatch
    #   typer
simple-websocket==1.1.0
    # via python-engineio
sniffio==1.3.1
    # via
    #   anyio
    #   asgi-lifespan
stack-data==0.6.3
    # via ipython
textual==3.3.0
    # via
    #   textual-dev
    #   textual-serve
textual-dev==1.7.0
    # via -r requirements-dev.in
textual-serve==1.1.2
    # via textual-dev
toml==0.10.2
    # via bunnet
tomli-w==1.2.0
    # via hatch
tomlkit==0.13.3
    # via hatch
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
trove-classifiers==2025.5.9.12
    # via hatchling
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   pyright
    #   rich-toolkit
    #   textual
    #   textual-dev
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.4.0
    # via
    #   geventhttpclient
    #   requests
userpath==1.9.2
    # via hatch
uv==0.7.12
    # via
    #   -r requirements-dev.in
    #   hatch
    #   pre-commit-uv
uvicorn==0.34.3
    # via fastapi-cli
uvloop==0.21.0
    # via uvicorn
virtualenv==20.31.2
    # via
    #   hatch
    #   pre-commit
watchfiles==1.0.5
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websocket-client==1.8.0
    # via python-socketio
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   flask-login
    #   locust
wsproto==1.2.0
    # via simple-websocket
yarl==1.20.0
    # via aiohttp
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent
zstandard==0.23.0
    # via hatch
